replaced grid propagator by python/numba functions, moved all GPU code to 32bit

+++ pisa/utils/const.py
@@ -1,2 +1,2 @@
 import numpy as np
-FTYPE = np.float64
+FTYPE = np.float32
